requirements/prod.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=prod.txt prod.txt.in
#
-e ..
alembic==1.2.1
argon2-cffi==19.2.0
attrs==19.3.0             # via jsonschema, openapi-core
certifi==2019.9.11        # via requests
cffi==1.13.1              # via argon2-cffi
chardet==3.0.4            # via requests
expandvars==0.4           # via pyramid-heroku
gunicorn==19.9.0
hupper==1.9               # via pyramid
idna==2.8                 # via requests
importlib-metadata==0.23  # via jsonschema
jsonschema==3.1.1         # via openapi-spec-validator
lazy-object-proxy==1.4.3  # via openapi-core
mako==1.1.0               # via alembic
markupsafe==1.1.1         # via mako
more-itertools==7.2.0     # via zipp
mypy-extensions==0.4.3
openapi-core==0.11.1      # via pyramid-openapi3
openapi-spec-validator==0.2.8  # via openapi-core, pyramid-openapi3
passlib==1.7.1
pastedeploy==2.0.1        # via plaster-pastedeploy
plaster-pastedeploy==0.7  # via pyramid
plaster==1.0              # via plaster-pastedeploy, pyramid
psycopg2-binary==2.8.4
pycparser==2.19           # via cffi
pyjwt==1.7.1              # via pyramid-jwt
pyramid-force-https==0.1.1
pyramid-heroku==0.6.0
pyramid-jwt==1.4.1
pyramid-openapi3==0.4.1
pyramid-redirect==0.3
pyramid-retry==2.1
pyramid-tm==2.3
pyramid==1.10.4
git+https://github.com/niteoweb/pyramid_deferred_sqla.git#egg=pyramid_deferred_sqla
pyrsistent==0.15.5        # via jsonschema
python-dateutil==2.8.0    # via alembic
python-editor==1.0.4      # via alembic
python-slugify==4.0.0
pyyaml==5.1.2             # via openapi-spec-validator
requests==2.22.0          # via pyramid-heroku
six==1.12.0               # via argon2-cffi, jsonschema, openapi-core, openapi-spec-validator, pyrsistent, python-dateutil, structlog
sqlalchemy==1.3.10        # via alembic, zope.sqlalchemy
strict-rfc3339==0.7       # via openapi-core
structlog==19.2.0
text-unidecode==1.3       # via python-slugify
transaction==2.4.0        # via pyramid-tm, zope.sqlalchemy
translationstring==1.3    # via pyramid
urllib3==1.25.6           # via requests
venusian==1.2.0           # via pyramid
webob==1.8.5              # via pyramid
zipp==0.6.0               # via importlib-metadata
zope.deprecation==4.4.0   # via pyramid
zope.interface==4.6.0     # via pyramid, pyramid-retry, transaction, zope.sqlalchemy
zope.sqlalchemy==1.2

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.6.0        # via jsonschema, plaster, pyramid, zope.deprecation, zope.interface, zope.sqlalchemy
